res comments
