+++ pkg/cmd/repo/rename/rename.go
@@ -1,15 +1,12 @@
 package rename
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/MakeNowJust/heredoc"
-	"github.com/cli/cli/v2/api"
 	"github.com/cli/cli/v2/context"
 	"github.com/cli/cli/v2/git"
 	"github.com/cli/cli/v2/internal/config"
@@ -30,13 +27,6 @@ type RenameOptions struct {
 	newRepoSelector string
 }
 
-type renameRepo struct {
-	RepoHost  string
-	RepoOwner string
-	RepoName  string
-	Name      string `json:"name,omitempty"`
-}
-
 func NewCmdRename(f *cmdutil.Factory, runf func(*RenameOptions) error) *cobra.Command {
 	opts := &RenameOptions{
 		IO:         f.IOStreams,
@@ -46,15 +36,11 @@ func NewCmdRename(f *cmdutil.Factory, runf func(*RenameOptions) error) *cobra.Co
 	}
 
 	cmd := &cobra.Command{
-		Use:   "rename [<repository>] [<new-name>]",
+		Use:   "rename [<new-name>]",
 		Short: "Rename a repository",
 		Long: heredoc.Doc(`Rename a GitHub repository
-		
-		With no argument, the repository for the current directory is renamed using a prompt
-		
-		With one argument, the repository of the current directory is renamed using the argument
-		
-		With '-R', and two arguments the given repository is replaced with the new name`),
+
+		By default, renames the current repository otherwise rename the specified repository.`),
 		Args: cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			opts.BaseRepo = f.BaseRepo
@@ -63,7 +49,7 @@ func NewCmdRename(f *cmdutil.Factory, runf func(*RenameOptions) error) *cobra.Co
 			if len(args) > 0 {
 				opts.newRepoSelector = args[0]
 			} else if !opts.IO.CanPrompt() {
-				return &cmdutil.FlagError{Err: errors.New("could not prompt: proceed with a repo name")}
+				return &cmdutil.FlagError{Err: errors.New("could not prompt: new name required when not running interactively")}
 			}
 
 			if runf != nil {
@@ -83,12 +69,9 @@ func renameRun(opts *RenameOptions) error {
 	if err != nil {
 		return err
 	}
-	apiClient := api.NewClientFromHTTP(httpClient)
 
-	var input renameRepo
 	var newRepo ghrepo.Interface
 	var baseRemote *context.Remote
-	var remoteUpdateError error
 	newRepoName := opts.newRepoSelector
 
 	currRepo, err := opts.BaseRepo()
@@ -99,7 +82,7 @@ func renameRun(opts *RenameOptions) error {
 	if newRepoName == "" {
 		err = prompt.SurveyAskOne(
 			&survey.Input{
-				Message: fmt.Sprintf("Rename %s to: ", currRepo.RepoOwner()+"/"+currRepo.RepoName()),
+				Message: fmt.Sprintf("Rename %s to: ", ghrepo.FullName(currRepo)),
 			},
 			&newRepoName,
 		)
@@ -108,21 +91,20 @@ func renameRun(opts *RenameOptions) error {
 		}
 	}
 
-	input = renameRepo{
-		RepoHost:  currRepo.RepoHost(),
-		RepoOwner: currRepo.RepoOwner(),
-		RepoName:  currRepo.RepoName(),
-		Name:      newRepoName,
-	}
-
 	newRepo = ghrepo.NewWithHost(currRepo.RepoOwner(), newRepoName, currRepo.RepoHost())
 
-	err = runRename(apiClient, currRepo.RepoHost(), input)
+	err = runRename(httpClient, currRepo, newRepoName)
 	if err != nil {
-		return fmt.Errorf("API called failed: %s, please check your parameters", err)
+		return fmt.Errorf("API called failed: %s", err)
+	}
+
+	if opts.IO.IsStdoutTTY() {
+		cs := opts.IO.ColorScheme()
+		fmt.Fprintf(opts.IO.Out, "%s Renamed repository %s\n", cs.SuccessIcon(), ghrepo.FullName(newRepo))
 	}
 
 	if !opts.HasRepoOverride {
+		cs := opts.IO.ColorScheme()
 		cfg, err := opts.Config()
 		if err != nil {
 			return err
@@ -132,30 +114,13 @@ func renameRun(opts *RenameOptions) error {
 		remotes, _ := opts.Remotes()
 		baseRemote, _ = remotes.FindByRepo(currRepo.RepoOwner(), currRepo.RepoName())
 		remoteURL := ghrepo.FormatRemoteURL(newRepo, protocol)
-		remoteUpdateError = git.UpdateRemoteURL(baseRemote.Name, remoteURL)
-		if remoteUpdateError != nil {
-			cs := opts.IO.ColorScheme()
+		err = git.UpdateRemoteURL(baseRemote.Name, remoteURL)
+		if err != nil {
 			fmt.Fprintf(opts.IO.ErrOut, "%s warning: unable to update remote '%s' \n", cs.WarningIcon(), err)
 		}
-	}
-
-	if opts.IO.IsStdoutTTY() {
-		cs := opts.IO.ColorScheme()
-		fmt.Fprintf(opts.IO.Out, "%s Renamed repository %s\n", cs.SuccessIcon(), input.RepoOwner+"/"+input.Name)
-		if !opts.HasRepoOverride && remoteUpdateError == nil {
+		if opts.IO.IsStdoutTTY() {
 			fmt.Fprintf(opts.IO.Out, "%s Updated the %q remote \n", cs.SuccessIcon(), baseRemote.Name)
 		}
 	}
 	return nil
-}
-
-func runRename(apiClient *api.Client, hostname string, input renameRepo) error {
-	path := fmt.Sprintf("repos/%s/%s", input.RepoOwner, input.RepoName)
-	body := &bytes.Buffer{}
-	enc := json.NewEncoder(body)
-	if err := enc.Encode(input); err != nil {
-		return err
-	}
-
-	return apiClient.REST(hostname, "PATCH", path, body, nil)
 }
